# users table: add css_id index