--- a/db/migrate/20210907174359_add_user_css_id_index.rb
+++ b/db/migrate/20210907174359_add_user_css_id_index.rb
@@ -1,0 +1,5 @@
+class AddUserCssIdIndex < Caseflow::Migration
+  def change
+    add_index(:users, :css_id, unique: true, algorithm: :concurrently)
+  end
+end