--- a/db/schema.rb
+++ b/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_09_01_121223) do
+ActiveRecord::Schema.define(version: 2021_09_07_174359) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -1535,6 +1535,7 @@ ActiveRecord::Schema.define(version: 2021_09_01_121223) do
     t.datetime "status_updated_at", comment: "When the user's status was last updated"
     t.datetime "updated_at"
     t.index "upper((css_id)::text)", name: "index_users_unique_css_id", unique: true
+    t.index ["css_id"], name: "index_users_on_css_id", unique: true
     t.index ["status"], name: "index_users_on_status"
     t.index ["updated_at"], name: "index_users_on_updated_at"
   end