Resolves these long-running queries: https://onenr.io/0z7wkn9plQL
```
SELECT "users".* FROM "users" WHERE "users"."css_id" IN ($?, $?, $?, $?, $?, $?, 
  $?, $?, $?, $?, $?, $?, $?, $?, $?, $?, $?, $?, $?, 
  $?, $?, $?, $?, $?, $?, $?, $?, $?, $?, $?, $?, $?, ...
```
https://onenr.io/0LkjnO99XQo

![image](https://user-images.githubusercontent.com/55255674/132389796-d5d88ed1-53e9-4e32-9b44-e8c39bfe0121.png)

[Slack](https://dsva.slack.com/archives/C2ZAMLK88/p1631037125070500?thread_ts=1631035856.066500&cid=C2ZAMLK88)

### Description
`users` table: add css_id index to make the above query efficient.

We added an `upper((css_id)::text))` index in PR #10188, but that won't help with the query above.
Rather than try to force searching by uppercase CSS_ID in all our code, let's just add an index.

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] Above query uses DB index

### Testing Plan
Before DB migration (note the `Seq Scan`):
```
User.where(css_id: User.first(10).pluck(:id)).explain
[2021-09-07 13:12:52 -0500]   User Load (1.4ms)  SELECT  "users".* FROM "users" ORDER BY "users"."id" ASC LIMIT $1  [["LIMIT", 10]]
[2021-09-07 13:12:52 -0500]   User Load (4.1ms)  SELECT "users".* FROM "users" WHERE "users"."css_id" IN ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)  [["css_id", "1"], ["css_id", "2"], ["css_id", "3"], ["css_id", "4"], ["css_id", "5"], ["css_id", "6"], ["css_id", "7"], ["css_id", "8"], ["css_id", "9"], ["css_id", "10"]]
=> EXPLAIN for: SELECT "users".* FROM "users" WHERE "users"."css_id" IN ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10) [["css_id", "1"], ["css_id", "2"], ["css_id", "3"], ["css_id", "4"], ["css_id", "5"], ["css_id", "6"], ["css_id", "7"], ["css_id", "8"], ["css_id", "9"], ["css_id", "10"]]
                             QUERY PLAN
---------------------------------------------------------------------
 Seq Scan on users  (cost=0.00..179.83 rows=10 width=175)
   Filter: ((css_id)::text = ANY ('{1,2,3,4,5,6,7,8,9,10}'::text[]))
```


After DB migration (note the `Index Scan`):
```
User.where(css_id: User.first(10).pluck(:id)).explain
[2021-09-07 13:07:32 -0500]   User Load (4.1ms)  SELECT  "users".* FROM "users" ORDER BY "users"."id" ASC LIMIT $1  [["LIMIT", 10]]
[2021-09-07 13:07:32 -0500]   User Load (1.9ms)  SELECT "users".* FROM "users" WHERE "users"."css_id" IN ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)  [["css_id", "1"], ["css_id", "2"], ["css_id", "3"], ["css_id", "4"], ["css_id", "5"], ["css_id", "6"], ["css_id", "7"], ["css_id", "8"], ["css_id", "9"], ["css_id", "10"]]
=> EXPLAIN for: SELECT "users".* FROM "users" WHERE "users"."css_id" IN ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10) [["css_id", "1"], ["css_id", "2"], ["css_id", "3"], ["css_id", "4"], ["css_id", "5"], ["css_id", "6"], ["css_id", "7"], ["css_id", "8"], ["css_id", "9"], ["css_id", "10"]]
                                      QUERY PLAN
--------------------------------------------------------------------------------------
 Bitmap Heap Scan on users  (cost=38.90..67.45 rows=10 width=175)
   Recheck Cond: ((css_id)::text = ANY ('{1,2,3,4,5,6,7,8,9,10}'::text[]))
   ->  Bitmap Index Scan on index_users_on_css_id  (cost=0.00..38.90 rows=10 width=0)
         Index Cond: ((css_id)::text = ANY ('{1,2,3,4,5,6,7,8,9,10}'::text[]))
```

### Database Changes
*Only for Schema Changes*

* [x] Have your migration classes inherit from `Caseflow::Migration`, especially when adding indexes (use `add_safe_index`)
* [x] Verify that `migrate:rollback` works as desired ([`change` supported functions](https://edgeguides.rubyonrails.org/active_record_migrations.html#using-the-change-method))
* [x] Perform query profiling (eyeball Rails log, check bullet and fasterer output)
* [x] Add appropriate indexes (especially for foreign keys, polymorphic columns, unique constraints, and Rails scopes)
* [x] Post this PR in #appeals-schema with a summary