@yoomlam 2021-09-07
Better:
![image](https://user-images.githubusercontent.com/55255674/132406116-3e2f0ffe-acba-4950-bab9-88ecc50c83c7.png)
At least it's using the index: https://onenr.io/0X8wo5JAvjx